Shipping helper/USPS/UPS dimensions and other fixes.
As promised, here are the modifications needed in order for the USPS
and UPS modules to get package dimensions to calculate shipping. These
dimensions were not set, so each module was rounding each package
dimension to 1 in. I added some filters and class properties to allow
for future shipment packaging optimization. If you require any
information, please don’t hesitate to ask.