Visitor Authentication Security Hardening
**Insipred by CVE-2014-0166 resolved in WordPress 3.8.2**
Instead of using wp_hash, a pluggable that can be overridden we use hash_hmac with a wll defined WPSC_SALT
Introduced WPSC_SALT an constant that can be defined in our config files, or takes on the value of the WordPress Authentication salt.
If the WordPress authentication salt is not available a unqiue salt is created using the file name, which should be cosntant, and the file path, which should vary from installation.to installation.

**Also...**
Removed a call to ctype, this library may not be available on all servers