Don't cut off checkout page when there are no quotes
There was a path through the checkout page where the forms will not be displayed.

When there aren't any valid shipping quotes, perhaps because configured shipping modules do not have enough information, the cehckout form will now be displayed.

I think it's easy enough to to fix, but we will need a release note so that support and users who encounter this issue will know to update their theme file.