Adjusted the check for when a thumbnail needs to be regenerated.
The previous check  appeared to be using the presence of a WPeC specific meta value as an indicator that an image of the correct proportions is available. If the meta value was not present then all of the image sizes for the object would be regenerated.  This is a time consuming and resource intensive operation. After the image was regenerated the WPeC specific meta value was not being set.  That means that the next time through the function the intermediate images sizes would be regenerated again, and again ....

There also appear to be frequent are conditions when the image of the correct size is available, but the meta value is not set.

The code was modified to have two distinct checks. Is the special meta key available, if not use the default value. Second check is if the image of the desired size, from the meta key or the default value for the meta key, is already available.