- abstracting gettin variation product id from variation selections
Here we pull out the loops in wpsc_add_to_cart that took the variation
selection and parsed it down to a variation_product_id and moved them in
to their own function.

This allows developers to leverage a core WPSC function instead of
maintaining their own loops to get the variation_product_id and then
needing to maintain them long term.

It also breaks out the huge function that is wpsc_add_to_cart in to one
more smaller part that we could write a unit test for without it being
buried deep down inside another function.