- passing back 2 arguements
Later on in the wpsc_add_to_cart function (line 151 on this PR) WPSC
merges the $default_parameters and the $provided_parameter arrays.

Then it checks for the $parameters['variation_values'] against ! is_null
on line 165 which would have been set in the previous loop that this PR
changes around.

To keep that from breaking this we may not see this passed back 2 items
in an array then reloads the $provided_parameters array with the
expected variation_values key.