In product grid, products with variations now have select options button
With variations and extra options being hidden in grid view, the add to cart
button could pose a lot of problems without selecting the appropriate choices.
Now when we are on pages with a body class of .wpsc-grid and the product has
variations, the add to cart button changes to select options button which
redirects the user to the single.php version of the product where they can choose
the needed variations for their product.